added delete document test case

+++ src/lib/deleteDocument.js
@@ -1,5 +1,5 @@
 //firebase imports
-import { collection, deleteDoc, doc } from "firebase/firestore";
+import { collection, deleteDoc, doc, getDoc } from "firebase/firestore";
 import { useCurrentUser } from "vuefire";
 import { db } from "@/main"; //firestore instance
 
@@ -10,5 +10,11 @@ export default async function deleteDocument(documentID) {
     collection(db, "users/" + userID + "/docs"),
     documentID
   );
-  await deleteDoc(currDocRef);
+
+  const docSnap = await getDoc(currDocRef);
+  if (docSnap.exists()) {
+    await deleteDoc(currDocRef);
+  } else {
+    throw new Error("Document does not exist");
+  }
 }
